+++ summingbird-chill/src/main/scala/com/twitter/summingbird/chill/SBChillRegistrar.scala
@@ -31,10 +31,10 @@ import com.twitter.summingbird.batch.{ BatchID, Timestamp }
 object SBChillRegistrar {
 
   def injectionRegistrar[T: Manifest: Codec]: InjectionRegistrar[T] =
-    InjectionRegistrar(manifest[T].erasure.asInstanceOf[Class[T]], implicitly[Codec[T]])
+    InjectionRegistrar(manifest[T].runtimeClass.asInstanceOf[Class[T]], implicitly[Codec[T]])
 
   def injectionDefaultRegistrar[T: Manifest: Codec]: InjectionDefaultRegistrar[T] =
-    InjectionDefaultRegistrar(manifest[T].erasure.asInstanceOf[Class[T]], implicitly[Codec[T]])
+    InjectionDefaultRegistrar(manifest[T].runtimeClass.asInstanceOf[Class[T]], implicitly[Codec[T]])
 
   def kryoRegClass(clazz: Class[_]*) =
     { k: Kryo =>